Rename `Secp256k1` + latest `pairing` lib traits
