+++ ecdsa/secp256k1/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "secp256k1forhalo2"
+name = "secp256k1"
 description = "Implementation of the secp256k1 with halo2 (or pallas curves) traits"
 version = "0.0.1"
 authors = [
@@ -37,7 +37,7 @@ harness = false
 
 [dependencies]
 
-maingate = { path = "../../maingate", default-features = false }
+halo2wrong = { path = "../../halo2wrong", default-features = false }
 blake2b_simd = { version = "0.5", default-features = false }
 ff = { version = "0.11", default-features = false }
 group = { version = "0.11", features = ["tests"] }
@@ -51,6 +51,6 @@ cfg-if = "0.1"
 
 [features]
 default = ["bits", "zcash"]
-kzg = ["maingate/kzg"]
-zcash = ["maingate/zcash"]
+zcash = ["halo2wrong/zcash"]
+kzg = ["halo2wrong/kzg"]
 bits = ["ff/bits"]
